add stub data and api endpoints
