+++ package.json
@@ -5,7 +5,7 @@
   "workspaces": {
     "packages": [
       "packages/nextjs",
-      "packages/local_db"
+      "packages/firebase"
     ]
   },
   "scripts": {
@@ -17,7 +17,7 @@
     "precommit": "lint-staged",
     "vercel": "yarn workspace @se-2/nextjs vercel",
     "vercel:yolo": "yarn workspace @se-2/nextjs vercel:yolo",
-    "seed": "yarn workspace local_db seed"
+    "seed": "yarn workspace firebase seed"
   },
   "packageManager": "yarn@3.2.3",
   "devDependencies": {
