+++ packages/firebase/.gitignore
@@ -1,0 +1,21 @@
+# dependencies
+/node_modules
+
+# build
+/dist
+
+# misc
+.DS_Store
+
+# local env files
+.env
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# typescript
+*.tsbuildinfo
+
+# seed 
+seed.json
