--- packages/local_db/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "local_db",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "seed": "tsx index.ts"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "dotenv": "^16.4.5",
-    "firebase-admin": "^11.11.1"
-  },
-  "devDependencies": {
-    "@types/node": "^20.12.12",
-    "tsx": "^4.11.2",
-    "typescript": "^5.4.5"
-  }
-}
