--- packages/local_db/seed.sample.json
@@ -1,277 +1,0 @@
-{
-  "version": "1.1.1",
-  "projects": {
-    "project-id1": {
-      "avatarUrl": "string",
-      "coverImageUrl": "string",
-      "attestationUid": "string",
-      "approvalAttestationUid": "string",
-      "name": "string",
-      "description": "string",
-      "externalLink": "string",
-      "socialLinks": {
-        "twitter": "string",
-        "farcaster": "string",
-        "mirror": "string",
-        "website": "string"
-      },
-      "team": [
-        {
-          "farcasterId": "string"
-        }
-      ],
-      "repositories": [
-        "string"
-      ],
-      "deployedContracts": [
-        {}
-      ],
-      "categories": [
-        {
-          "name": "string",
-          "description": "string"
-        }
-      ],
-      "funding": {
-        "ventureCapital": [
-          {
-            "amount": "string",
-            "source": "string",
-            "date": "2024-06-07T18:41:57.328Z",
-            "details": "string"
-          }
-        ],
-        "grants": [
-          {
-            "amount": "string",
-            "source": "string",
-            "date": "2024-06-07T18:41:57.328Z",
-            "details": "string"
-          }
-        ],
-        "optimismGrants": [
-          {}
-        ]
-      }
-    },
-    "project-id2": {
-      "avatarUrl": "string",
-      "coverImageUrl": "string",
-      "attestationUid": "string",
-      "approvalAttestationUid": "string",
-      "name": "string",
-      "description": "string",
-      "externalLink": "string",
-      "socialLinks": {
-        "twitter": "string",
-        "farcaster": "string",
-        "mirror": "string",
-        "website": "string"
-      },
-      "team": [
-        {
-          "farcasterId": "string"
-        }
-      ],
-      "repositories": [
-        "string"
-      ],
-      "deployedContracts": [
-        {}
-      ],
-      "categories": [
-        {
-          "name": "string",
-          "description": "string"
-        }
-      ],
-      "funding": {
-        "ventureCapital": [
-          {
-            "amount": "string",
-            "source": "string",
-            "date": "2024-06-07T18:41:57.328Z",
-            "details": "string"
-          }
-        ],
-        "grants": [
-          {
-            "amount": "string",
-            "source": "string",
-            "date": "2024-06-07T18:41:57.328Z",
-            "details": "string"
-          }
-        ],
-        "optimismGrants": [
-          {}
-        ]
-      }
-    }
-  },
-  "metrics": {
-    "onchain-users": {
-      "name": "string",
-      "description": "string",
-      "weight": 1
-    },
-    "tx-fees": {
-      "name": "string",
-      "description": "string",
-      "weight": 1
-    }
-  },
-  "totals-by-date":{
-    "20240101": {
-      "totalProjects": 157,
-      "totalOverallScore": 15898,
-      "metrics": {
-        "onchain-users": 4078,
-        "tx-fees": 11820
-      }
-    }
-  },
-  "project-totals-by-date": {
-    "20240101": {
-      "project-id1": {
-        "isNew": false,
-        "overallScore": 1202,
-        "rank": 1,
-        "rankChange": 0,
-        "rankChangeWeek": 0,
-        "rankChangeMonth": 1,
-        "momentum": 2,
-        "metrics": {
-          "onchain-users": {
-            "score": 1001,
-            "rank": 1,
-            "momentum": 2
-          },
-          "tx-fees": {
-            "score": 101,
-            "rank": 2,
-            "momentum": 2
-          }
-        }
-      },
-      "project-id2": {
-        "isNew": false,
-        "overallScore": 1202,
-        "rank": 2,
-        "rankChange": 0,
-        "rankChangeWeek": 0,
-        "rankChangeMonth": -1,
-        "momentum": 2,
-        "metrics": {
-          "onchain-users": {
-            "score": 1001,
-            "rank": 1,
-            "momentum": 2
-          },
-          "tx-fees": {
-            "score": 101,
-            "rank": 2,
-            "momentum": 2
-          }
-        }
-      }
-    }
-  },
-  "etl-process-log": {
-    "20240101": {
-      "status": "success",
-      "message": "string"
-    }
-  },
-  "raw-oso-response": {
-    "20240101": [
-      {
-        "active_contract_count_90_days": 1,
-        "address_count": 2,
-        "address_count_90_days": 2,
-        "days_since_first_transaction": 201,
-        "display_name": "Daimo",
-        "event_source": "OPTIMISM",
-        "gas_fees_sum": 0.00082130956988073,
-        "gas_fees_sum_6_months": 0.0008183154836908259,
-        "high_activity_address_count_90_days": 0,
-        "low_activity_address_count_90_days": 1,
-        "medium_activity_address_count_90_days": 1,
-        "multi_project_address_count_90_days": 0,
-        "new_address_count_90_days": 0,
-        "project_id": "6RbJ7fhcEoin39N2pN8WMvaRQe3TVP_JaohQ7nvojCs=",
-        "project_name": "daimo-eth",
-        "project_namespace": "oso",
-        "project_source": "OSS_DIRECTORY",
-        "returning_address_count_90_days": 2,
-        "transaction_count": 34,
-        "transaction_count_6_months": 30
-      },
-      {
-        "active_contract_count_90_days": 1,
-        "address_count": 2,
-        "address_count_90_days": 2,
-        "days_since_first_transaction": 242,
-        "display_name": "Bankless-Africa",
-        "event_source": "OPTIMISM",
-        "gas_fees_sum": 0.0005384550051846021,
-        "gas_fees_sum_6_months": 0.000526745507020252,
-        "high_activity_address_count_90_days": 0,
-        "low_activity_address_count_90_days": 1,
-        "medium_activity_address_count_90_days": 1,
-        "multi_project_address_count_90_days": 2,
-        "new_address_count_90_days": 0,
-        "project_id": "TnnHcvj2jZDYBPeLzwAMUzlynJygKL7dJnP2imJzQYk=",
-        "project_name": "bankless-africa",
-        "project_namespace": "oso",
-        "project_source": "OSS_DIRECTORY",
-        "returning_address_count_90_days": 2,
-        "transaction_count": 62,
-        "transaction_count_6_months": 54
-      },
-      {
-        "active_contract_count_90_days": 1,
-        "address_count": 3,
-        "address_count_90_days": 2,
-        "days_since_first_transaction": 200,
-        "display_name": "Endangered Tokens: Endangered Trees as ReFi Biodiversity Assets ",
-        "event_source": "OPTIMISM",
-        "gas_fees_sum": 0.000031688086604034,
-        "gas_fees_sum_6_months": 0.000021458559807634,
-        "high_activity_address_count_90_days": 0,
-        "low_activity_address_count_90_days": 2,
-        "medium_activity_address_count_90_days": 0,
-        "multi_project_address_count_90_days": 0,
-        "new_address_count_90_days": 1,
-        "project_id": "NEpTyRk2iJgqJmeLr4-u4hX2hgLeR38Qn3oUZuk1SVs=",
-        "project_name": "endangeredtokens",
-        "project_namespace": "oso",
-        "project_source": "OSS_DIRECTORY",
-        "returning_address_count_90_days": 1,
-        "transaction_count": 22,
-        "transaction_count_6_months": 12
-      },
-      {
-        "active_contract_count_90_days": 2,
-        "address_count": 359,
-        "address_count_90_days": 2,
-        "days_since_first_transaction": 714,
-        "display_name": "Manifold",
-        "event_source": "OPTIMISM",
-        "gas_fees_sum": 0.10770561886011508,
-        "gas_fees_sum_6_months": 0.00043673927702894997,
-        "high_activity_address_count_90_days": 0,
-        "low_activity_address_count_90_days": 2,
-        "medium_activity_address_count_90_days": 0,
-        "multi_project_address_count_90_days": 0,
-        "new_address_count_90_days": 2,
-        "project_id": "a5hVSzyhsfqMcoWljR1DzjmwLFlDEhnU08YVni8xrPw=",
-        "project_name": "manifold",
-        "project_namespace": "oso",
-        "project_source": "OSS_DIRECTORY",
-        "returning_address_count_90_days": 0,
-        "transaction_count": 1012,
-        "transaction_count_6_months": 6
-      }
-    ]
-  }
-}
