--- packages/local_db/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "noUncheckedIndexedAccess": true,
-    "skipLibCheck": true,
-    "target": "es2022",
-    "allowJs": true,
-    "resolveJsonModule": true,
-    "moduleDetection": "force",
-    "isolatedModules": true,
-    "moduleResolution": "NodeNext",
-    "esModuleInterop": true,
-    "module": "NodeNext",
-    "outDir": "dist"
-  }
-}
